+++ x/rollapp/genesisbridge/genesis_transfer.go
@@ -1,4 +1,4 @@
-package transfergenesis
+package genesisbridge
 
 import (
 	"encoding/json"
