+++ x/rollapp/genesisbridge/memo_test.go
@@ -1,4 +1,4 @@
-package transfergenesis
+package genesisbridge
 
 import (
 	"testing"
