+++ x/rollapp/genesisbridge/testdata/memo_happy_path_with_other_namespace.json
