+++ x/rollapp/genesisbridge/testdata/memo_namespace_empty.json
